requirements.txt
lxml==5.1.0
python-docx==1.1.0
typing_extensions==4.9.0
